consensus: move tendermint types to tendermint validator
